+++ tests/test_pubnet.py
@@ -133,30 +133,8 @@ class TestNodes:
 
 
 class TestNetwork:
-    @pytest.mark.filterwarnings(
-        "ignore:Constructing PubNet object without Publication nodes."
-    )
-    def test_handles_no_nodes(self):
-        net = PubNet.load_graph(
-            "simple_pubnet",
-            None,
-            (("Publication", "Author"),),
-            data_dir="tests/data",
-        )
-        assert len(net["Publication"]) == 0
-        assert len(net["Author"]) == 0
-
     def test_creates_empty_nodes_for_missing_edge_nodes(self, simple_pubnet):
         assert len(simple_pubnet["Chemical"]) == 0
-
-    def test_handles_no_edges(self):
-        net = PubNet.load_graph(
-            "simple_pubnet",
-            ("Publication",),
-            None,
-            data_dir="tests/data",
-        )
-        assert len(net.edges) == 0
 
     def test_filter_to_single_publication_id(self, simple_pubnet):
         publication_id = 1